# go/docs/swagger.yaml
basePath: /
definitions:
  auth.AuthResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
  auth.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  auth.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: password123
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  httputil.APIResponse:
    properties:
      code:
        example: ENTRY_CREATED
        type: string
      correlationId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      data: {}
      error:
        example: INVALID_REQUEST
        type: string
      message:
        example: Request processed successfully
        type: string
      responseTime:
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
...
  models.UserResponse:
    properties:
      email:
        example: user@example.com
        type: string
      id:
        example: 507f1f77bcf86cd799439011
        type: string
      name:
        example: John Doe
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@dict-simulator.io
    name: API Support
  description: A simulated implementation of the Brazilian Central Bank's DICT API
    for managing Pix keys.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DICT Simulator API
  version: 1.0.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user with email and password. Returns a JWT token
        on success.
      parameters:
      - description: User login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/httputil.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.AuthResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/httputil.APIResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/httputil.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/httputil.APIResponse'
      summary: User login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email, password, and name. Returns
        a JWT token on success.
      parameters:
      - description: User registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            allOf:
            - $ref: '#/definitions/httputil.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.AuthResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/httputil.APIResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/httputil.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/httputil.APIResponse'
      summary: Register a new user
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"